schemes_rag_agent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pinecone>=2.2.0
sentence-transformers>=2.2.0
pymongo>=4.6.0
httpx>=0.25.0
python-multipart>=0.0.6
psutil>=5.9.0

# A2A Protocol Dependencies
a2a-python>=0.1.0
starlette>=0.27.0
